Fixed a warning on Start()
Although this is actually a symptom of a larger issue. Because PlayerClass extends EntityClass, PlayerClass itself has a "healthbar" which should be set to prevent a null reference. However, evidentially, playerclass is meant to be an intermediate class inbetween the concrete individual as well as providing functionality for card drawing. Therefore, this Start() should be a override of the actual EntityClass , Jackie should extend playerClass and would be overriding this Start()