completely implemented Linear Search and LIFO Invariant and Player Priority and "Speed Exclusion"; Data required for testing in description.
Linear Search is completely implemented; I noticed that it doesn't need to return the index of a FOUND card since that behaviour is taken care of inside Dequeue.

LIFO insertion: I believe this is redundant since a call to PlayerAction INserting and Enemy action invoke the same method that ensures LIFO is always maintained with the player being given priority; this needs to be tested: need a new player and enemy actions and possibly more cards (for speed comparison)
Have pre-empted insertion of a same speed card by a character; this too needs to be tested by both an additional character and more cards. (you can find this code inside the BQ class; felt more appropriate to just do it there).