Added a background fade to better highlight the characters
Did some really sus UI stuff cuz i simply dont understand unity UI :(
Forcing the canvases to be a certain way for health bar and buff bar. Changed the sorting order of everything else to be consistent so that it can be faded properly.

Changed a z value to reduce weird overlaps