Almost* done
There's a bug in either of InsertIntoWrapperArray or SortWrappers; I think it is the latter as I feel uncertain about the logic. you use while for this sort. Have implemented dequeue but can't test until I have Sort correctly implemented. Finally there's a print method that prints the wrappers to the array. Not too bad just need to debug.