Completed Wrapper Implemention
Read Title.
Qualifier:
2 things remain:
Initiating the EnemyActionTransformation() based on GameState (Easy Fix but need to be careful so as to not cause TWO transformations)
Thereafter the rendering of a new BattleQueue. I say thereafter because the BQ is rendered BEFORE the player has added ANY action which could result in a rather weird "transition".