Implemented HipFire,  Rapid Fire, Flurry
Flurry needs to be tested completely: it belongs to Ive; but it should be fine since the logic resembles others.

HipFire was very hard to test and kept throwing errors. Have implemented the logic but everything else (like TextInfo? remains)

Similar with RapidFire's Duplicate card. I do not know how to populate the fields inside the class. Why should they not be auto populated?

Note: there are TWO prefabs for RapidFire one in newPrefabs and one in an other folder v confusing. To  resolve.